+++ src/ts/Test/Unit/AdUnits/AdUnitFactoryTest.ts
@@ -29,6 +29,7 @@ import { IAdUnitParameters } from 'AdUnits/AbstractAdUnit';
 import { Campaign } from 'Models/Campaign';
 
 import ConfigurationJson from 'json/ConfigurationAuctionPlc.json';
+import { ConfigurationParser } from 'Parsers/ConfigurationParser';
 import { MoatViewabilityService } from 'Utilities/MoatViewabilityService';
 import { XPromoAdUnit } from 'AdUnits/XPromoAdUnit';
 import { XPromoCampaign } from 'Models/Campaigns/XPromoCampaign';
@@ -36,6 +37,7 @@ import { PromoCampaign } from 'Models/Campaigns/PromoCampaign';
 import { PromoAdUnit } from 'AdUnits/PromoAdUnit';
 import { PurchasingUtilities } from 'Utilities/PurchasingUtilities';
 import { OperativeEventManagerFactory } from 'Managers/OperativeEventManagerFactory';
+import { GdprConsentManager } from 'Managers/GdprConsentManager';
 
 describe('AdUnitFactoryTest', () => {
 
@@ -67,11 +69,12 @@ describe('AdUnitFactoryTest', () => {
         sandbox.stub(container, 'close').returns(Promise.resolve());
         sandbox.stub(container, 'open').returns(Promise.resolve());
         thirdPartyEventManager = new ThirdPartyEventManager(nativeBridge, request);
-        config = new Configuration(JSON.parse(ConfigurationJson));
+        config = ConfigurationParser.parse(JSON.parse(ConfigurationJson));
         deviceInfo = <DeviceInfo>{getLanguage: () => 'en', getAdvertisingIdentifier: () => '000', getLimitAdTracking: () => false, getOsVersion: () => '8.0'};
         clientInfo = TestFixtures.getClientInfo(Platform.ANDROID);
         sessionManager = new SessionManager(nativeBridge, request);
         const campaign = TestFixtures.getCampaign();
+        const gdprManager = sinon.createStubInstance(GdprConsentManager);
 
         operativeEventManager = OperativeEventManagerFactory.createOperativeEventManager({
             nativeBridge: nativeBridge,
@@ -98,7 +101,8 @@ describe('AdUnitFactoryTest', () => {
             campaign: campaign,
             configuration: config,
             request: request,
-            options: {}
+            options: {},
+            gdprManager: gdprManager
         };
 
         sandbox.spy(thirdPartyEventManager, 'sendEvent');