+++ src/ts/Test/Unit/EventHandlers/PromoEventHandlerTest.ts
@@ -45,7 +45,7 @@ describe('PromoEventHandlersTest', () => {
         it('should hide adunit', () => {
             promoAdUnit = sinon.createStubInstance(PromoAdUnit);
 
-            PromoEventHandler.onClose(nativeBridge, promoAdUnit, '111', '111', 1, [purchaseTrackingUrls]);
+            PromoEventHandler.onClose(nativeBridge, promoAdUnit, '111', '111', 1, [purchaseTrackingUrls], false);
             sinon.assert.called(<sinon.SinonSpy>promoAdUnit.hide);
         });
     });
@@ -54,7 +54,7 @@ describe('PromoEventHandlersTest', () => {
         it('should hide adunit', () => {
             promoAdUnit = sinon.createStubInstance(PromoAdUnit);
 
-            PromoEventHandler.onClose(nativeBridge, promoAdUnit, '111', '111', 1, [purchaseTrackingUrls]);
+            PromoEventHandler.onClose(nativeBridge, promoAdUnit, '111', '111', 1, [purchaseTrackingUrls], false);
             sinon.assert.called(<sinon.SinonSpy>promoAdUnit.hide);
         });
 
@@ -67,6 +67,26 @@ describe('PromoEventHandlersTest', () => {
 
             PromoEventHandler.onPromo(nativeBridge, promoAdUnit, 'com.unit.test.iapproductid', [purchaseTrackingUrls]);
             sinon.assert.called(<sinon.SinonSpy>PurchasingUtilities.beginPurchaseEvent);
+        });
+    });
+
+    describe('when calling onGDPRPopupSkipped', () => {
+        const operativeEventManager = sinon.createStubInstance(OperativeEventManager);
+
+        it ('should set the optOutRecorded flag in the configuration', () => {
+            const config = sinon.createStubInstance(Configuration);
+
+            (<sinon.SinonStub>config.isOptOutRecorded).returns(false);
+
+            PromoEventHandler.onGDPRPopupSkipped(config, operativeEventManager);
+            sinon.assert.called(<sinon.SinonSpy>config.setOptOutRecorded);
+        });
+
+        it('should send GDPR operative Event with skip', () => {
+            const config = sinon.createStubInstance(Configuration);
+
+            PromoEventHandler.onGDPRPopupSkipped(config, operativeEventManager);
+            sinon.assert.calledWith(<sinon.SinonSpy>operativeEventManager.sendGDPREvent, 'skip');
         });
     });
 });