+++ src/ts/Test/Unit/EventHandlers/VastOverlayEventHandlersTest.ts
@@ -26,6 +26,7 @@ import { MOAT } from 'Views/MOAT';
 import { MoatViewabilityService } from 'Utilities/MoatViewabilityService';
 import { SinonSandbox } from 'sinon';
 import { OperativeEventManagerFactory } from 'Managers/OperativeEventManagerFactory';
+import { GdprConsentManager } from 'Managers/GdprConsentManager';
 
 describe('VastOverlayEventHandlersTest', () => {
     let campaign: VastCampaign;
@@ -100,6 +101,8 @@ describe('VastOverlayEventHandlersTest', () => {
             campaign: campaign
         });
 
+        const gdprManager = sinon.createStubInstance(GdprConsentManager);
+
         vastAdUnitParameters = {
             forceOrientation: Orientation.LANDSCAPE,
             focusManager: focusManager,
@@ -115,7 +118,8 @@ describe('VastOverlayEventHandlersTest', () => {
             options: {},
             endScreen: undefined,
             overlay: overlay,
-            video: campaign.getVideo()
+            video: campaign.getVideo(),
+            gdprManager: gdprManager
         };
 
         vastAdUnit = new VastAdUnit(nativeBridge, vastAdUnitParameters);
@@ -208,7 +212,7 @@ describe('VastOverlayEventHandlersTest', () => {
 
         it('should call video click through tracking url', () => {
             sinon.stub(nativeBridge, 'getPlatform').returns(Platform.IOS);
-            sinon.stub(nativeBridge.UrlScheme, 'open');
+            sinon.stub(nativeBridge.UrlScheme, 'open').resolves();
             vastOverlayEventHandler.onOverlayCallButton().then(() => {
                 sinon.assert.calledOnce(<sinon.SinonSpy>vastAdUnit.sendVideoClickTrackingEvent);
             });
@@ -216,7 +220,7 @@ describe('VastOverlayEventHandlersTest', () => {
 
         it('should open click trough link in iOS web browser when call button is clicked', () => {
             sinon.stub(nativeBridge, 'getPlatform').returns(Platform.IOS);
-            sinon.stub(nativeBridge.UrlScheme, 'open');
+            sinon.stub(nativeBridge.UrlScheme, 'open').resolves();
             vastOverlayEventHandler.onOverlayCallButton().then(() => {
                 sinon.assert.calledWith(<sinon.SinonSpy>nativeBridge.UrlScheme.open, 'http://foo.com');
             });
@@ -224,7 +228,7 @@ describe('VastOverlayEventHandlersTest', () => {
 
         it('should open click trough link in Android web browser when call button is clicked', () => {
             sinon.stub(nativeBridge, 'getPlatform').returns(Platform.ANDROID);
-            sinon.stub(nativeBridge.Intent, 'launch');
+            sinon.stub(nativeBridge.Intent, 'launch').resolves();
             vastOverlayEventHandler.onOverlayCallButton().then(() => {
                 sinon.assert.calledWith(<sinon.SinonSpy>nativeBridge.Intent.launch, {
                     'action': 'android.intent.action.VIEW',