+++ src/ts/Utilities/CustomFeatures.ts
@@ -1,4 +1,3 @@
-
 import { AdUnitStyle } from 'Models/AdUnitStyle';
 import { StorageType } from 'Native/Api/Storage';
 import { NativeBridge } from 'Native/NativeBridge';
@@ -24,6 +23,10 @@ export class CustomFeatures {
 
     public static isFancyEndScreenEnabled(abGroup: number): boolean {
         // TODO: Use actual group(s) when we know them
-        return abGroup === 17;
+        return abGroup === 5 || abGroup === 6;
+    }
+
+    public static isGDPRBaseTest(abGroup: number): boolean {
+        return abGroup === 16 || abGroup === 17;
     }
 }