+++ src/ts/Views/EndScreen.ts
@@ -1,5 +1,4 @@
 import EndScreenTemplate from 'html/EndScreen.html';
-import SquareEndScreenTemplate from 'html/SquareEndScreen.html';
 import FancyEndScreenTemplate from 'html/FancyEndScreen.html';
 
 import { NativeBridge } from 'Native/NativeBridge';
@@ -12,7 +11,6 @@ import { Campaign } from 'Models/Campaign';
 import { IEndScreenDownloadParameters } from 'EventHandlers/EndScreenEventHandler';
 import { AdUnitStyle } from 'Models/AdUnitStyle';
 import { CustomFeatures } from 'Utilities/CustomFeatures';
-import { SquareEndScreenUtilities } from 'Utilities/SquareEndScreenUtilities';
 import { Platform } from 'Constants/Platform';
 
 export interface IEndScreenHandler {
@@ -22,9 +20,8 @@ export interface IEndScreenHandler {
     onGDPRPopupSkipped(): void;
 }
 
-const GDPR_OPT_OUT_BASE  = 'gdpr-pop-up-base';
-
-const SQUARE_END_SCREEN = 'square-end-screen';
+const GDPR_OPT_OUT_BASE = 'gdpr-pop-up-base';
+const GDPR_OPT_OUT_ICON = 'gdpr-pop-up-icon';
 
 const FANCY_END_SCREEN = 'fancy-end-screen';
 
@@ -60,7 +57,7 @@ export abstract class EndScreen extends View<IEndScreenHandler> implements IPriv
             {
                 event: 'click',
                 listener: (event: Event) => this.onDownloadEvent(event),
-                selector: '.game-background, .download-container, .game-icon, .game-image'
+                selector: '.game-background, .download-container, .game-icon'
             },
             {
                 event: 'click',
@@ -74,7 +71,7 @@ export abstract class EndScreen extends View<IEndScreenHandler> implements IPriv
             },
             {
                 event: 'click',
-                listener: (event: Event) => this.onGDPRPopupEvent(event),
+                listener: (event: Event) => this.onPrivacyEvent(event),
                 selector: '.gdpr-link'
             }
         ];
@@ -110,7 +107,16 @@ export abstract class EndScreen extends View<IEndScreenHandler> implements IPriv
         const endScreenAlt = this.getEndscreenAlt();
         if (typeof endScreenAlt === 'string') {
             this._container.classList.add(endScreenAlt);
-            document.documentElement.classList.add(endScreenAlt);
+
+            /* If pop up is visible, hide privacy button */
+            if (endScreenAlt === GDPR_OPT_OUT_BASE || endScreenAlt === GDPR_OPT_OUT_ICON) {
+                (<HTMLElement>this._container.querySelector('.privacy-button')).style.display = 'none';
+            }
+        }
+
+        /* TODO: Should go away once we finish with a/b test */
+        if (!CustomFeatures.isGDPRBaseTest(this._abGroup)) {
+            this._container.classList.add('use-gdpr-privacy-icon');
         }
     }
 
@@ -163,13 +169,7 @@ export abstract class EndScreen extends View<IEndScreenHandler> implements IPriv
 
     protected getEndscreenAlt(campaign?: Campaign) {
         if (this._showGDPRBanner) {
-            return GDPR_OPT_OUT_BASE;
-        }
-
-        const campaignId = campaign ? campaign.getId() : this._campaignId;
-        const platform = this._nativeBridge.getPlatform();
-        if (SquareEndScreenUtilities.useSquareEndScreenAlt(this._abGroup, platform, campaignId, this._osVersion)) {
-            return SQUARE_END_SCREEN;
+            return CustomFeatures.isGDPRBaseTest(this._abGroup) ? GDPR_OPT_OUT_BASE : GDPR_OPT_OUT_ICON;
         }
 
         if (CustomFeatures.isFancyEndScreenEnabled(this._abGroup) && this.canShowFancyEndScreen()) {
@@ -189,12 +189,9 @@ export abstract class EndScreen extends View<IEndScreenHandler> implements IPriv
     private onPrivacyEvent(event: Event): void {
         event.preventDefault();
 
-        this._privacy.show();
-    }
-
-    private onGDPRPopupEvent(event: Event) {
-        event.preventDefault();
-        this._gdprPopupClicked = true;
+        if (this._showGDPRBanner) {
+            this._gdprPopupClicked = true;
+        }
         this._privacy.show();
     }
 
@@ -207,9 +204,7 @@ export abstract class EndScreen extends View<IEndScreenHandler> implements IPriv
     }
 
     private getTemplate() {
-        if (this.getEndscreenAlt() === SQUARE_END_SCREEN) {
-            return SquareEndScreenTemplate;
-        }  else if (this.getEndscreenAlt() === FANCY_END_SCREEN) {
+        if (this.getEndscreenAlt() === FANCY_END_SCREEN) {
             return FancyEndScreenTemplate;
         }